implement session manager and memcache manager crete a data package to store all clases definitions closes #1
