+++ registerUsageLog.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+import sys
+from data.usagelog import usageLog
+import session_manager
+
+
+usageType = sys.argv[1]
+
+usage = usageLog(type=usageType)
+session = session_manager.getSession()
+session.add(usage)
+session.commit()
+
